WIP, WARNING (precision loss): vectorize dalpha updates.
Since dalpha calculations affect the alpha calculations, which affect the z0 calculations which affect ... so on and so forth, the small numerical differences introduced through this vectorization (maybe due to floating point associativity) propogate/ripple through several other variables. Thus, for several variables I had to lower the precision by 1 order of magnitude (e.g. 6 decimals to 5 decimals). For a few select variables, the absolute differences raises substantially (e.g. 5 to 50).

Most importantly, we lost precision on the alpha matrix (at the end of the program) when compared to the fortran output, by 1 order of magnitude (3 decimals to 2 decimals).

Also importantly, our Log likelihood precision is not affected, and the correlations between the fortran and python outputs remain above .99

i Please enter the commit message for your changes. Lines starting