WIP, Vectorize rho updates (kind of)... to be continued
np.vectorize is just a python loop in a wrapper and doesnt give us any optimization.

Really we should use scipy.special.digamma instead of psifun (which is amica's custom implementation) but lets wait until later when most of our tests our stripped out and we can just impact the effect of it on the final results.