Remove loglik from AmicaState
AmicaState should contain only the persistent parameters required to use the model (W, A, mu, sbeta, etc.). You could save this single object, load it a year later, and apply it to new data to get the unmixed sources. It is the essential, data-independent result of the fitting process.

loglik should go in `AmicaMetrics`, the Story of the Training. It contains the ephemeral, per-iteration diagnostics that tell you how the model was fit (iter, loglik, lrate, gradient_norm).