Insantiate AmicaUpdates in the local scope.
The arrays stored in AmicaUpdates are relatively small (<1mb) so allocating them on each iteration should  not have to big an impact, and it makes the get_updates.. function more pure.

I left all the machinery for resetting updates in place, so that we can reert this commit later if we find that pre allocating is more optimal. We can keep the tests in the local scope if we do.