Pre allocate AmicaUpdates (Once again)
The undoes the work of previous commit which moved the allocation of Updates within the tight loop.

There is a case to be made against such allocation in a tight loop, even if the arrays are really small here.

But really, get_updates_and.. now takes in AmicaMetrics, modifies it and returns it, so let's just make it behave the same with AmicaUpdates (consume it, update it, return it) to make the behavior consistent.