Dont (potentially) overwrite into alpha or sbeta array
alpha_br and sbeta_br happen to be copies because they were created with advanced indexing.

If that ever changes (e.g. they are created via simple indexing or slicing etc) then now alpha_br becomes a view, and overwiting it also overwrites alpha, which we do not want!

So dont depend on this brittle scenario..