WIP: Fix test
Due to numerical instability across runs SOMETIMES we get a division by zero error in the simulated data. So I put a catch for this that throws a warnign so we still know it happened.

In the future we may want to move this regularization to the function that computes y - but for now let's be very cautious and conservative to not perturb the path away from the Fortran implementaiton by default